use throttle in useResponsive hook and refactor
