+++ src/shared/ui/checkbox/checkbox.tsx
@@ -2,8 +2,7 @@
 import React, { ComponentPropsWithoutRef, ElementRef, ReactNode, forwardRef, useId } from 'react'
 
 import CheckIcon from '@/shared/assets/icons/CheckIcon'
-import { EMPTY_STRING } from '@/shared/constants'
-import { ReturnComponent } from '@/shared/types'
+import { EMPTY_STRING, ReturnComponent } from '@/shared'
 import * as CheckboxRadix from '@radix-ui/react-checkbox'
 import * as LabelRadix from '@radix-ui/react-label'
 import clsx from 'clsx'
